+++ x/finality/keeper/power_table_test.go
@@ -8,9 +8,11 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 
+	testutil "github.com/babylonlabs-io/babylon/testutil/btcstaking-helper"
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
 	btclctypes "github.com/babylonlabs-io/babylon/x/btclightclient/types"
 	"github.com/babylonlabs-io/babylon/x/btcstaking/types"
+	ftypes "github.com/babylonlabs-io/babylon/x/finality/types"
 )
 
 func FuzzVotingPowerTable(f *testing.F) {
@@ -24,21 +26,20 @@ func FuzzVotingPowerTable(f *testing.F) {
 		// mock BTC light client and BTC checkpoint modules
 		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
 		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		finalityKeeper := types.NewMockFinalityKeeper(ctrl)
-		finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), gomock.Any(), gomock.Any()).Return(true).AnyTimes()
-		h := NewHelper(t, btclcKeeper, btccKeeper, finalityKeeper)
+		h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
 
 		// set all parameters
 		covenantSKs, _ := h.GenAndApplyParams(r)
 		changeAddress, err := datagen.GenRandomBTCAddress(r, h.Net)
 		h.NoError(err)
 
-		// generate a random batch of finality providers
+		// generate a random batch of finality providers, and commit pub rand list with timestamp
 		fps := []*types.FinalityProvider{}
 		numFpsWithVotingPower := datagen.RandomInt(r, 10) + 2
 		numFps := numFpsWithVotingPower + datagen.RandomInt(r, 10)
 		for i := uint64(0); i < numFps; i++ {
-			_, _, fp := h.CreateFinalityProvider(r)
+			fpSK, _, fp := h.CreateFinalityProvider(r)
+			h.CommitPubRandList(r, fpSK, fp, 1, 100, true)
 			fps = append(fps, fp)
 		}
 
@@ -74,25 +75,27 @@ func FuzzVotingPowerTable(f *testing.F) {
 		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
 		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
 		require.NoError(t, err)
+		err = h.FinalityKeeper.BeginBlocker(h.Ctx)
+		require.NoError(t, err)
 
 		for i := uint64(0); i < numFpsWithVotingPower; i++ {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
 			require.Equal(t, numBTCDels*stakingValue, power)
 		}
 		for i := numFpsWithVotingPower; i < numFps; i++ {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
 			require.Zero(t, power)
 		}
 
 		// also, get voting power table and assert consistency
-		powerTable := h.BTCStakingKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
+		powerTable := h.FinalityKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
 		require.NotNil(t, powerTable)
 		for i := uint64(0); i < numFpsWithVotingPower; i++ {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
 			require.Equal(t, powerTable[fps[i].BtcPk.MarshalHex()], power)
 		}
 		// the activation height should be the current Babylon height as well
-		activatedHeight, err := h.BTCStakingKeeper.GetBTCStakingActivatedHeight(h.Ctx)
+		activatedHeight, err := h.FinalityKeeper.GetBTCStakingActivatedHeight(h.Ctx)
 		require.NoError(t, err)
 		require.Equal(t, babylonHeight, activatedHeight)
 
@@ -113,10 +116,12 @@ func FuzzVotingPowerTable(f *testing.F) {
 		// index height and record power table
 		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
 		require.NoError(t, err)
+		err = h.FinalityKeeper.BeginBlocker(h.Ctx)
+		require.NoError(t, err)
 
 		// check if the slashed finality provider's voting power becomes zero
 		for i := uint64(0); i < numFpsWithVotingPower; i++ {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
 			if i == slashedIdx {
 				require.Zero(t, power)
 			} else {
@@ -124,15 +129,15 @@ func FuzzVotingPowerTable(f *testing.F) {
 			}
 		}
 		for i := numFpsWithVotingPower; i < numFps; i++ {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
 			require.Zero(t, power)
 		}
 
 		// also, get voting power table and assert consistency
-		powerTable = h.BTCStakingKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
+		powerTable = h.FinalityKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
 		require.NotNil(t, powerTable)
 		for i := uint64(0); i < numFpsWithVotingPower; i++ {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
 			if i == slashedIdx {
 				require.Zero(t, power)
 			}
@@ -150,14 +155,94 @@ func FuzzVotingPowerTable(f *testing.F) {
 		require.NoError(t, err)
 
 		for _, fp := range fps {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fp.BtcPk, babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fp.BtcPk, babylonHeight)
 			require.Zero(t, power)
 		}
 
 		// the activation height should be same as before
-		activatedHeight2, err := h.BTCStakingKeeper.GetBTCStakingActivatedHeight(h.Ctx)
+		activatedHeight2, err := h.FinalityKeeper.GetBTCStakingActivatedHeight(h.Ctx)
 		require.NoError(t, err)
 		require.Equal(t, activatedHeight, activatedHeight2)
+	})
+}
+func FuzzRecordVotingPowerDistCache(f *testing.F) {
+	datagen.AddRandomSeedsToFuzzer(f, 10)
+
+	f.Fuzz(func(t *testing.T, seed int64) {
+		r := rand.New(rand.NewSource(seed))
+		ctrl := gomock.NewController(t)
+		defer ctrl.Finish()
+
+		// mock BTC light client and BTC checkpoint modules
+		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
+		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
+		h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
+
+		// set all parameters
+		covenantSKs, _ := h.GenAndApplyParams(r)
+		changeAddress, err := datagen.GenRandomBTCAddress(r, h.Net)
+		h.NoError(err)
+
+		// generate a random batch of finality providers, and commit
+		// pub rand list with timestamp
+		numFpsWithVotingPower := datagen.RandomInt(r, 10) + 2
+		numFps := numFpsWithVotingPower + datagen.RandomInt(r, 10)
+		fpsWithVotingPowerMap := map[string]*types.FinalityProvider{}
+		for i := uint64(0); i < numFps; i++ {
+			fpSK, _, fp := h.CreateFinalityProvider(r)
+			h.CommitPubRandList(r, fpSK, fp, 1, 100, true)
+			if i < numFpsWithVotingPower {
+				// these finality providers will receive BTC delegations and have voting power
+				fpsWithVotingPowerMap[fp.Addr] = fp
+			}
+		}
+
+		// for the first numFpsWithVotingPower finality providers, generate a random number of BTC
+		// delegations and add covenant signatures to activate them
+		numBTCDels := datagen.RandomInt(r, 10) + 1
+		stakingValue := datagen.RandomInt(r, 100000) + 100000
+		for _, fp := range fpsWithVotingPowerMap {
+			for j := uint64(0); j < numBTCDels; j++ {
+				delSK, _, err := datagen.GenRandomBTCKeyPair(r)
+				h.NoError(err)
+				stakingTxHash, delMsg, del, btcHeaderInfo, inclusionProof, _, err := h.CreateDelegation(
+					r,
+					delSK,
+					fp.BtcPk.MustToBTCPK(),
+					changeAddress.EncodeAddress(),
+					int64(stakingValue),
+					1000,
+					0,
+					0,
+					true,
+				)
+				h.NoError(err)
+				h.CreateCovenantSigs(r, covenantSKs, delMsg, del)
+				h.AddInclusionProof(stakingTxHash, btcHeaderInfo, inclusionProof)
+			}
+		}
+
+		// record voting power distribution cache
+		babylonHeight := datagen.RandomInt(r, 10) + 1
+		h.Ctx = datagen.WithCtxHeight(h.Ctx, babylonHeight)
+		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
+		h.BeginBlocker()
+
+		// assert voting power distribution cache is correct
+		dc := h.FinalityKeeper.GetVotingPowerDistCache(h.Ctx, babylonHeight)
+		require.NotNil(t, dc)
+		require.Equal(t, dc.TotalBondedSat, numFpsWithVotingPower*numBTCDels*stakingValue, dc.String())
+		activeFPs := dc.GetActiveFinalityProviderSet()
+		for _, fpDistInfo := range activeFPs {
+			require.Equal(t, fpDistInfo.TotalBondedSat, numBTCDels*stakingValue)
+			fp, ok := fpsWithVotingPowerMap[fpDistInfo.Addr]
+			require.True(t, ok)
+			require.Equal(t, fpDistInfo.Commission, fp.Commission)
+			require.Len(t, fpDistInfo.BtcDels, int(numBTCDels))
+			for _, delDistInfo := range fpDistInfo.BtcDels {
+				require.Equal(t, delDistInfo.TotalSat, stakingValue)
+			}
+		}
 	})
 }
 
@@ -172,8 +257,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 		// mock BTC light client and BTC checkpoint modules
 		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
 		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		finalityKeeper := types.NewMockFinalityKeeper(ctrl)
-		h := NewHelper(t, btclcKeeper, btccKeeper, finalityKeeper)
+		h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
 
 		// set all parameters
 		covenantSKs, _ := h.GenAndApplyParams(r)
@@ -181,12 +265,12 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 		h.NoError(err)
 
 		// generate a random batch of finality providers, each with a BTC delegation with random power
-		fpsWithMeta := []*types.FinalityProviderDistInfo{}
+		fpsWithMeta := []*ftypes.FinalityProviderDistInfo{}
 		numFps := datagen.RandomInt(r, 300) + 1
 		noTimestampedFps := map[string]bool{}
 		for i := uint64(0); i < numFps; i++ {
 			// generate finality provider
-			_, _, fp := h.CreateFinalityProvider(r)
+			fpSK, _, fp := h.CreateFinalityProvider(r)
 
 			// delegate to this finality provider
 			stakingValue := datagen.RandomInt(r, 100000) + 100000
@@ -209,13 +293,13 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 
 			// 30 percent not have timestamped randomness, which causes
 			// zero voting power in the table
-			fpDistInfo := &types.FinalityProviderDistInfo{BtcPk: fp.BtcPk, TotalBondedSat: stakingValue}
+			fpDistInfo := &ftypes.FinalityProviderDistInfo{BtcPk: fp.BtcPk, TotalBondedSat: stakingValue}
 			if r.Intn(10) <= 2 {
-				finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), fp.BtcPk, gomock.Any()).Return(false).AnyTimes()
+				h.CommitPubRandList(r, fpSK, fp, 1, 100, false)
 				noTimestampedFps[fp.BtcPk.MarshalHex()] = true
 				fpDistInfo.IsTimestamped = false
 			} else {
-				finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), fp.BtcPk, gomock.Any()).Return(true).AnyTimes()
+				h.CommitPubRandList(r, fpSK, fp, 1, 100, true)
 				fpDistInfo.IsTimestamped = true
 			}
 
@@ -223,9 +307,9 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 			fpsWithMeta = append(fpsWithMeta, fpDistInfo)
 		}
 
-		maxActiveFpsParam := h.BTCStakingKeeper.GetParams(h.Ctx).MaxActiveFinalityProviders
+		maxActiveFpsParam := h.FinalityKeeper.GetParams(h.Ctx).MaxActiveFinalityProviders
 		// get a map of expected active finality providers
-		types.SortFinalityProvidersWithZeroedVotingPower(fpsWithMeta)
+		ftypes.SortFinalityProvidersWithZeroedVotingPower(fpsWithMeta)
 		expectedActiveFps := fpsWithMeta[:min(uint32(len(fpsWithMeta)-len(noTimestampedFps)), maxActiveFpsParam)]
 		expectedActiveFpsMap := map[string]uint64{}
 		for _, fp := range expectedActiveFps {
@@ -236,12 +320,11 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 		babylonHeight := datagen.RandomInt(r, 10) + 1
 		h.SetCtxHeight(babylonHeight)
 		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
-		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
-		require.NoError(t, err)
+		h.BeginBlocker()
 
 		// only finality providers in expectedActiveFpsMap have voting power
 		for _, fp := range fpsWithMeta {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, fp.BtcPk.MustMarshal(), babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, fp.BtcPk.MustMarshal(), babylonHeight)
 			if expectedPower, ok := expectedActiveFpsMap[fp.BtcPk.MarshalHex()]; ok {
 				require.Equal(t, expectedPower, power)
 			} else {
@@ -251,7 +334,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 
 		// also, get voting power table and assert there is
 		// min(len(expectedActiveFps), MaxActiveFinalityProviders) active finality providers
-		powerTable := h.BTCStakingKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
+		powerTable := h.FinalityKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
 		expectedNumActiveFps := len(expectedActiveFpsMap)
 		if expectedNumActiveFps > int(maxActiveFpsParam) {
 			expectedNumActiveFps = int(maxActiveFpsParam)
@@ -275,25 +358,23 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 		// mock BTC light client and BTC checkpoint modules
 		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
 		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		finalityKeeper := types.NewMockFinalityKeeper(ctrl)
-		finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), gomock.Any(), gomock.Any()).Return(true).AnyTimes()
-		h := NewHelper(t, btclcKeeper, btccKeeper, finalityKeeper)
+		h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
 
 		// set all parameters
 		covenantSKs, _ := h.GenAndApplyParams(r)
 		// set random number of max number of finality providers
 		// in order to cover cases that number of finality providers is more or
 		// less than `MaxActiveFinalityProviders`
-		bsParams := h.BTCStakingKeeper.GetParams(h.Ctx)
-		bsParams.MaxActiveFinalityProviders = uint32(datagen.RandomInt(r, 20) + 10)
-		err := h.BTCStakingKeeper.SetParams(h.Ctx, bsParams)
+		fParams := h.FinalityKeeper.GetParams(h.Ctx)
+		fParams.MaxActiveFinalityProviders = uint32(datagen.RandomInt(r, 20) + 10)
+		err := h.FinalityKeeper.SetParams(h.Ctx, fParams)
 		h.NoError(err)
 		// change address
 		changeAddress, err := datagen.GenRandomBTCAddress(r, h.Net)
 		h.NoError(err)
 
 		numFps := datagen.RandomInt(r, 20) + 10
-		numActiveFPs := int(min(numFps, uint64(bsParams.MaxActiveFinalityProviders)))
+		numActiveFPs := int(min(numFps, uint64(fParams.MaxActiveFinalityProviders)))
 
 		/*
 			Generate a random batch of finality providers, each with a BTC delegation
@@ -304,7 +385,8 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 		for i := uint64(0); i < numFps; i++ {
 			// generate finality provider
 			// generate and insert new finality provider
-			_, fpPK, fp := h.CreateFinalityProvider(r)
+			fpSK, fpPK, fp := h.CreateFinalityProvider(r)
+			h.CommitPubRandList(r, fpSK, fp, 1, 100, true)
 
 			// create BTC delegation and add covenant signatures to activate it
 			stakingValue := datagen.RandomInt(r, 100000) + 100000
@@ -336,19 +418,18 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 		babylonHeight := datagen.RandomInt(r, 10) + 1
 		h.Ctx = datagen.WithCtxHeight(h.Ctx, babylonHeight)
 		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
-		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
-		h.NoError(err)
+		h.BeginBlocker()
 
 		// assert that only top `min(MaxActiveFinalityProviders, numFPs)` finality providers have voting power
 		sort.SliceStable(fpsWithMeta, func(i, j int) bool {
 			return fpsWithMeta[i].VotingPower > fpsWithMeta[j].VotingPower
 		})
 		for i := 0; i < numActiveFPs; i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Equal(t, fpsWithMeta[i].VotingPower, votingPower)
 		}
 		for i := numActiveFPs; i < int(numFps); i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Zero(t, votingPower)
 		}
 
@@ -390,11 +471,12 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 		// create more finality providers
 		numNewFps := datagen.RandomInt(r, 20) + 10
 		numFps += numNewFps
-		numActiveFPs = int(min(numFps, uint64(bsParams.MaxActiveFinalityProviders)))
+		numActiveFPs = int(min(numFps, uint64(fParams.MaxActiveFinalityProviders)))
 		for i := uint64(0); i < numNewFps; i++ {
 			// generate finality provider
 			// generate and insert new finality provider
-			_, fpPK, fp := h.CreateFinalityProvider(r)
+			fpSK, fpPK, fp := h.CreateFinalityProvider(r)
+			h.CommitPubRandList(r, fpSK, fp, 1, 100, true)
 
 			// create BTC delegation and add covenant signatures to activate it
 			stakingValue := datagen.RandomInt(r, 100000) + 100000
@@ -426,19 +508,18 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 		babylonHeight += 1
 		h.Ctx = datagen.WithCtxHeight(h.Ctx, babylonHeight)
 		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
-		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
-		h.NoError(err)
+		h.BeginBlocker()
 
 		// again, assert that only top `min(MaxActiveFinalityProviders, numFPs)` finality providers have voting power
 		sort.SliceStable(fpsWithMeta, func(i, j int) bool {
 			return fpsWithMeta[i].VotingPower > fpsWithMeta[j].VotingPower
 		})
 		for i := 0; i < numActiveFPs; i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Equal(t, fpsWithMeta[i].VotingPower, votingPower)
 		}
 		for i := numActiveFPs; i < int(numFps); i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Zero(t, votingPower)
 		}
 	})